refactor: helper func & types
