+++ packages/core/src/types.ts
@@ -31,7 +31,3 @@ export type SecretInjectionStrategy =
   | ({ kind: 'envFrom'; containerIndex?: number } & BaseSecretInjectionStrategy)
   | { kind: 'plugin'; action?: string; args?: unknown[]; [key: string]: unknown };
 
-export type StackFactory<Input, ResourceMap> = {
-  name: string;
-  create: (input: Input) => ResourceMap;
-};
